Complete exercises 4.15-4.23 Added token authentication for uses, tests for blogs, tests for users and more middleware to handle requests

+++ controllers/blogs.js
@@ -1,8 +1,10 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
@@ -16,9 +18,24 @@ blogsRouter.get('/:id', async (request, response) => {
 })
 
 blogsRouter.post('/', async (request, response) => {
-  const blog = new Blog(request.body)
-  const result = await blog.save()
-  response.status(201).json(result)
+  const body = request.body
+  const userId = request.user._id.toString()
+
+  const user = await User.findById(userId)
+
+  const blog = new Blog({
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes,
+    user: user._id
+  })
+
+  const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
+  response.json(savedBlog)
 })
 
 blogsRouter.put('/:id', async (request, response) => {
@@ -33,6 +50,11 @@ blogsRouter.put('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const blogToDelete = await Blog.findById(request.params.id)
+
+  if (blogToDelete.user.toString() !== request.user._id.toString()) {
+    return response.status(401).send({ error: 'only owner of blog is able to delete' })
+  }
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
